Refactor styles and update configuration
- Removed old CSS file (index.css) and replaced it with a new consolidated styles.css file containing global styles, typography, buttons, and layout components.
- Updated main.tsx to import the new styles.css instead of the old index.css.
- Enhanced TypeScript configuration files (tsconfig.app.json, tsconfig.json, tsconfig.node.json) for better module resolution and linting options.
- Modified vite.config.ts to use the @vitejs/plugin-react-swc for React support with Preact, including server and build configurations.
