+++ .npmrc
@@ -1,0 +1,5 @@
+hoist=false
+auto-install-peers=true
+dedupe-peer-dependents=false
+resolution-mode=highest
+enable-pre-post-scripts=true
