+++ package.json
@@ -5,15 +5,17 @@
     "type": "module",
     "scripts": {
         "dev": "vite",
-        "build": "tsc -b && vite build",
-        "preview": "vite preview"
+        "build": "vite build",
+        "preview": "vite preview",
+        "lint": "oxlint"
     },
     "dependencies": {
         "preact": "^10.27.0"
     },
     "devDependencies": {
-        "@preact/preset-vite": "^2.10.2",
-        "typescript": "~5.8.3",
-        "vite": "^7.1.2"
+        "@vitejs/plugin-react-swc": "^4.0.1",
+        "oxlint": "^1.13.0",
+        "typescript": "~5.9.2",
+        "vite": "^7.1.3"
     }
 }
